# src/main/java/co/edu/uniquindio/proyecto/dto/response/ErrorResponse.java
package co.edu.uniquindio.proyecto.dto.response;

import java.time.LocalDateTime;

public record ErrorResponse(
        LocalDateTime timestamp,
        String message,
        String code,
        String path) {
}